fix: wait for webhook listener completion
@octokit/webhooks.js is mostly tailored toward a callback function
world (it's particularly suited toward Express.js, etc.) this makes
things a little more complicated in the Promised land. In particular,
the Lambda could potentially exit early with asynchronous actions still
in progress, due to the fact we had nothing to let us guarantee
completion of the function passed through to `webhooks.on`. To solve it,
we wrap `webhooks.on` in a Promise, and use the `resolve`/`reject`
callbacks when we're ready, and just return this promise at the end.