perf: do label syncing asynchronously
GitHub Webhooks expect a response within 10 seconds, but the various
mutations to deal with Labels in GraphQL are fairly slow (~120ms each)
which will result in timeouts for any org with a decent number of labels
and/or repos. So deferring the real work to a separate Lambda and giving
an HTTP 204 seems safer.