fix: prevent surplus label deletion/creation/edit
GitHub will send a webhook every time a label is created, edited, or
deleted. This is no exception for when the app itself does its job,
which can potentially mean lots of webhooks. To avoid a whole heap of
redundant attempts at managing labels due to these extra requests, a
lock table can be used to bail out early in the event a label is already
midway through being managed.

This doesn't completely solve the problem, as GitHub will tend to
stagger webhooks apart at some rate, which can be after we've released
the lock. In this situation, we'll still talk to the GitHub API to see
what repos need to have labels created/edited/deleted, but because the
original transaction is complete, GitHub will come back saying there's
no labels/repos needing work, and we just exit. Ideally, we track more
of this in state so that we don't have to wait for the overhead of
talking to the GitHub API (which can also burn through a fair bit of
GitHub API quota).

Additionally label edits may still have this issue apply for now, as we
simply grab every label with a matching name (or old name, in the case
of a name edit), and just tell every one of them to update to the new
properties. I.e. we don't filter out labels where all their properties
already match. We just update all labels where the name would suggest
it's a candidate for an edit.